Add Primitive Header Parsing
